Добавлена функция определения даты, времени и региона, где было проведено сканирование
