+++ main.py
@@ -1,10 +1,6 @@
 #!/usr/bin/python3
 
-from scapy.all import *
-import argparse
-import datetime
-import pytz
-import geocoder
+#from scapy.all import *
 
 from tcp_ACK_scan import *
 from tcp_CON_scan import *
@@ -14,15 +10,8 @@ from sctp_INIT_scan import *
 from sctp_COOKIE_scan import *
 from mac import *
 from params import *
+from date_reg import *  
 
-
-def date_and_time():
-    location = geocoder.ip('me')
-    city = location.city
-    tz = pytz.timezone(location.timezone)
-    current_time = datetime.datetime.now(tz)
-    format_time = current_time.strftime("%d-%m-%Y %H: %M %Z")
-    print(f"Сканирование начато в {format_time} {city}")   
 
 def parse_ports(port_arg):
 
@@ -56,6 +45,8 @@ def main():
         return
 
     target_ports = parse_ports(args.ports)
+    
+    date_and_time()
 
     for port in target_ports:
         if args.tcp_syn_scan:
